% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/budgets_operations.R
\name{budgets_describe_budget}
\alias{budgets_describe_budget}
\title{Describes a budget}
\usage{
budgets_describe_budget(AccountId, BudgetName)
}
\arguments{
\item{AccountId}{[required] The \code{accountId} that is associated with the budget that you want a
description of.}

\item{BudgetName}{[required] The name of the budget that you want a description of.}
}
\description{
Describes a budget.
}
\section{Request syntax}{
\preformatted{budgets$describe_budget(
  AccountId = "string",
  BudgetName = "string"
)
}
}

\keyword{internal}
